Makefile now ignores tufle local files
The tufte classes can be customized with help of tufte-common-local.tex,
tufte-book-local.tex, and tufte-handout-local.tex files. The makefile as
it is written would try to compile these files into pdf, which would most
likely fail. This change changes makefile so it ignores these files.

Cleaned up the way those files are loaded in each class so it is more
standardized.