Fix the raggedouter setting for captions
The command that sets text alignment of captions was not working
properly in cases where twopage option was on and caption was set
to raggedouter. It uses the @tufte@float@recto, but it seems to not
work as intended and text was always ragged right. This change fixes
that issue.

This makes the \@tufte@justification@caption@outer and
\@tufte@justification@outer identical, hence they can be merged into
one if desired, this requires change on line 250 to the second command.